Login System Should work
DO NOT USE THIS IN PRODUCTION.
While this is for a school project we are not making a refresh token. Typically you would do this so people cant steal your access token since expires after one hour. We cannot manually destory JWTs so when we "logout" we simply stop the user from verifying with out session.

For our case this is fine. but proactively using this in the real world would NOT work as intended and our project would be open to attacks.